battlemaps/requirements.txt
asgiref==3.8.1
Django==4.2.18
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
psycopg==3.2.3
pytz==2024.2
sqlparse==0.5.3
